fix(CkCueEditor): extract FCk_CueToolbox_CueInfo to CkCueToolbox_Common.h to avoid unity-build ODR collision
Toolbox/CkCueToolbox.h (Slate consumer) and
UtilityWidget/CkCueToolbox_EditorUtilityWidget.h (UMG/Blueprint
consumer) each defined a struct named FCk_CueToolbox_CueInfo with
identical field layout but different reflection annotations — plain
C++ struct in the Slate header, USTRUCT(BlueprintType) in the UMG
header. Per-file builds kept them in separate TUs; unity merged them
and the duplicate type definitions collided.

Renaming per file doesn't work for duplicate type definitions (call
sites need to see matching types), so this module deviates from the
rename-per-file strategy used by CkInput/CkPmg/CkStateMachine/
CkAssetExporter and instead collapses the two definitions into one.

Extract the USTRUCT version (the more capable one, since Blueprint
reflection is a superset of what Slate needs) to a new header
Toolbox/CkCueToolbox_Common.h, matching the framework's existing
*_Common.h naming convention for shared type definitions
(CkOverlapBody_Common.h, CkPhysics_Common.h, etc.). Both consumer
headers now #include it instead of defining the struct themselves.

Consumers of the struct are unchanged — Slate code treats USTRUCTs as
normal C++ types with extra reflection metadata attached, so nothing
at the call-site level needs to change.

With the collision fixed, flip CkCueEditor.Build.cs back to the
default (unity-on) base constructor, reverting the opt-out from
de816388.
