+++ Source/CkCueEditor/CkCueEditor.Build.cs
@@ -3,7 +3,7 @@ using UnrealBuildTool;
 
 public class CkCueEditor : CkModuleRules
 {
-    public CkCueEditor(ReadOnlyTargetRules Target) : base(Target, UseUnityBuild: false)
+    public CkCueEditor(ReadOnlyTargetRules Target) : base(Target)
     {
         PrivateIncludePaths.AddRange(new string[] {
             // ... add other private include paths required here ...
